[ServiceBus] update release date (#23893)

+++ sdk/servicebus/azure-servicebus/azure/servicebus/_servicebus_client.py
@@ -170,7 +170,7 @@ class ServiceBusClient(object):
         *,
         retry_total: int = 3,
         retry_backoff_factor: float = 0.8,
-        retry_backoff_max: int = 120,
+        retry_backoff_max: float = 120,
         retry_mode: str = "exponential",
         **kwargs: Any
     ) -> "ServiceBusClient":
